requirements.txt
boto3==1.28.78
ask-sdk-core==1.19.0
ask-sdk-s3-persistence-adapter==1.0.0
ask-sdk-dynamodb-persistence-adapter==1.19.0
backports.zoneinfo==0.2.1
tzdata==2024.1

# FastAPI web app / Dialogflow webhook
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
jinja2>=3.1.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
